Move the connect dialog back into the integrations page
Keeps the picker's markup next to the page that owns it rather than in a
component module of its own. Behavior is unchanged: the catalog logic stays
in lib/preset-catalog.ts, and ConnectIntegrationDialog is file-local again.
